+++ client/app/models/reviewer-report.js
@@ -5,5 +5,8 @@ export default NestedQuestionOwner.extend({
   decision: DS.belongsTo('decision'),
   task: DS.belongsTo('task'),
   user: DS.belongsTo('user'),
+  status: DS.attr('string'),
+  statusDate: DS.attr('date'),
+  revision: DS.attr('string'),
   createdAt: DS.attr('date')
 });