+++ client/app/router.js
@@ -37,8 +37,16 @@ Router.map(function() {
   this.route('profile', { path: '/profile' });
 
   this.route('admin', function() {
-    this.route('journals', function() {});
+    this.route('cc', function() {
+      this.route('journals', function() {
+        this.route('cards');
+        this.route('workflows');
+        this.route('users');
+        this.route('settings');
+      });
+    });
 
+    this.route('journals', function() {});
     this.route('journal', { path: '/journals/:journal_id' }, function() {
       this.route('manuscript_manager_template', { path: '/manuscript_manager_templates' }, function() {
         this.route('new');
@@ -46,7 +54,7 @@ Router.map(function() {
       });
 
     });
-    this.route('feature-flags');
+    this.route('feature_flags');
   });
 });