+++ engines/tahi_standard_tasks/app/controllers/tahi_standard_tasks/upload_manuscript_controller.rb
@@ -18,7 +18,6 @@ module TahiStandardTasks
         params[:url],
         current_user
       )
-
       head 204
     end