+++ engines/tahi_standard_tasks/app/subscribers/paper/data_extracted/finish_upload_manuscript_task.rb
@@ -3,9 +3,9 @@ class Paper::DataExtracted::FinishUploadManuscriptTask
     return unless event_data[:record].completed?
     paper = Paper.find(event_data[:record].paper_id)
     unless paper.file_type == 'pdf'
-      paper
-      .tasks_for_type(TahiStandardTasks::UploadManuscriptTask.to_s)
-      .each(&:complete!)
+      paper.tasks_for_type(TahiStandardTasks::UploadManuscriptTask.to_s).each do |task|
+        task.complete!
+        Activity.task_updated! task, user: User.find(event_data[:record].user_id)
     end
   end
 end