+++ CHANGELOG.md
@@ -4,9 +4,11 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 
 ## ## [x.x.x] - {yyyy-mm-dd}
 ### Added
-* Users can now change the initial participants on a discussion topic
+* Users can now change the initial participants on a discussion topic.
 * The manuscript upload card prompts for a backing source file if a PDF manuscript is uploaded.
+* Users can now change the initial participants on a discussion topic.
 ### Changed
+* The Upload Manuscript card will uncheck as incomplete when a revision related decision is registered
 ### Deprecated
 ### Removed
 ### Fixed
@@ -26,6 +28,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ### Added
 * Logging outbound email sends to database, including status and forensics, to troubleshoot silent failures
 * Aperta can now use the title, abstract, and body HTML extracted from PDFs.
+* Display review status at the top of the task
 ### Changed
 * Users can download previous versions of a Manuscript
 ### Deprecated